backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
requests==2.32.5
pillow==11.3.0
